Moved mock data to data_loaders.tests
ECOM-5094
