+++ course_discovery/apps/course_metadata/data_loaders/tests/test_api.py
@@ -11,12 +11,11 @@ from course_discovery.apps.core.tests.utils import mock_api_callback, mock_jpeg_
 from course_discovery.apps.course_metadata.data_loaders.api import (
     OrganizationsApiDataLoader, CoursesApiDataLoader, EcommerceApiDataLoader, AbstractDataLoader, ProgramsApiDataLoader
 )
-from course_discovery.apps.course_metadata.data_loaders.tests import JSON, JPEG
+from course_discovery.apps.course_metadata.data_loaders.tests import JSON, JPEG, mock_data
 from course_discovery.apps.course_metadata.data_loaders.tests.mixins import ApiClientTestMixin, DataLoaderTestMixin
 from course_discovery.apps.course_metadata.models import (
     Course, CourseRun, Organization, Seat, Program, ProgramType,
 )
-from course_discovery.apps.course_metadata.tests import mock_data
 from course_discovery.apps.course_metadata.tests.factories import (
     CourseRunFactory, SeatFactory, ImageFactory, VideoFactory, OrganizationFactory, CourseFactory,
 )
